Intent: Strengthen cache modularity for coherent utilities Operators involved: [Resonance|Self-organization] Affected invariants: [#7, #8] Key changes: - moved cache helpers into tnfr.utils.cache and re-exported through utils - updated imports across engine, benchmarks, and tests to new module path - added tnfr.cache deprecation shim to maintain compatibility with warnings Expected risks/dissonances: Deprecation warning may surface new import sites; validated core cache tests Metrics: pytest tests/test_cache_helpers.py tests/test_stable_json.py Equivalence map: tnfr.cache.* -> tnfr.utils.cache.*
