+++ src/tnfr/dynamics/sampling.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from ..cache import cached_node_list
+from ..utils.cache import cached_node_list
 from ..rng import _rng_for_step, base_seed
 
 __all__ = ("update_node_sample",)
@@ -13,7 +13,7 @@ def update_node_sample(G, *, step: int) -> None:
     simulation step. When the network is small (``< 50`` nodes) or the limit
     is non‑positive, the full node set is used and sampling is effectively
     disabled. A snapshot of nodes is cached via a
-    :class:`~tnfr.cache.NodeCache` instance stored in
+    :class:`~tnfr.utils.cache.NodeCache` instance stored in
     ``G.graph['_node_list_cache']`` and reused across steps; it is only refreshed
     when the graph size changes. Sampling operates directly on the cached
     tuple of nodes.
