+++ src/tnfr/metrics/coherence.py
@@ -15,7 +15,7 @@ from ..callback_utils import CallbackEvent, callback_manager
 from ..glyph_history import ensure_history, append_metric
 from ..alias import collect_attr, get_attr, set_attr
 from ..helpers.numeric import clamp01
-from ..cache import ensure_node_index_map
+from ..utils.cache import ensure_node_index_map
 from .common import compute_coherence, min_max_range
 from .trig_cache import compute_theta_trig, get_trig_cache
 from ..observers import (
