+++ src/tnfr/metrics/trig_cache.py
@@ -11,7 +11,7 @@ from dataclasses import dataclass
 from typing import Any, Iterable, Mapping
 
 from ..alias import get_attr
-from ..cache import edge_version_cache
+from ..utils.cache import edge_version_cache
 from ..constants import get_aliases
 from ..types import GraphLike
 from ..utils import get_numpy
