+++ src/tnfr/utils/__init__.py
@@ -22,6 +22,23 @@ from .graph import (
     mark_dnfr_prep_dirty,
     supports_add_edge,
 )
+from .cache import (
+    EdgeCacheManager,
+    LockAwareLRUCache,
+    NODE_SET_CHECKSUM_KEY,
+    cached_node_list,
+    cached_nodes_and_A,
+    clear_node_repr_cache,
+    edge_version_cache,
+    edge_version_update,
+    ensure_node_index_map,
+    ensure_node_offset_map,
+    get_graph_version,
+    increment_edge_version,
+    increment_graph_version,
+    node_set_checksum,
+    stable_json,
+)
 from .io import (
     DEFAULT_PARAMS,
     JsonDumpsParams,
@@ -49,6 +66,21 @@ __all__ = (
     "MAX_MATERIALIZE_DEFAULT",
     "negative_weights_warn_once",
     "mix_groups",
+    "EdgeCacheManager",
+    "LockAwareLRUCache",
+    "NODE_SET_CHECKSUM_KEY",
+    "cached_node_list",
+    "cached_nodes_and_A",
+    "clear_node_repr_cache",
+    "edge_version_cache",
+    "edge_version_update",
+    "ensure_node_index_map",
+    "ensure_node_offset_map",
+    "get_graph_version",
+    "increment_edge_version",
+    "increment_graph_version",
+    "node_set_checksum",
+    "stable_json",
     "get_graph",
     "get_graph_mapping",
     "mark_dnfr_prep_dirty",
